requirements.txt
requests
discord.py
pyyaml